Extract and rename post-renderers to filters for clarity
